Move subscriber to it's own to keep actions isolated
